refactor: Question Controller/ManageService 생성 및 관련 메서드 이동
