+++ backend/bang-ggood/src/test/java/com/bang_ggood/checklist/service/ChecklistManageServiceTest.java
@@ -3,16 +3,6 @@ package com.bang_ggood.checklist.service;
 import com.bang_ggood.IntegrationTestSupport;
 import com.bang_ggood.checklist.ChecklistFixture;
 import com.bang_ggood.checklist.dto.request.ChecklistRequest;
-import com.bang_ggood.global.exception.BangggoodException;
-import com.bang_ggood.global.exception.ExceptionCode;
-import com.bang_ggood.question.CustomChecklistFixture;
-import com.bang_ggood.question.domain.CustomChecklistQuestion;
-import com.bang_ggood.question.domain.Question;
-import com.bang_ggood.question.dto.request.CustomChecklistUpdateRequest;
-import com.bang_ggood.question.dto.response.CategoryQuestionsResponse;
-import com.bang_ggood.question.dto.response.CustomChecklistQuestionsResponse;
-import com.bang_ggood.question.dto.response.QuestionResponse;
-import com.bang_ggood.question.repository.CustomChecklistQuestionRepository;
 import com.bang_ggood.user.UserFixture;
 import com.bang_ggood.user.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
@@ -20,25 +10,12 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import java.util.Collection;
-import java.util.List;
-
-import static com.bang_ggood.question.CustomChecklistFixture.CUSTOM_CHECKLIST_QUESTION_DEFAULT;
-import static com.bang_ggood.question.CustomChecklistFixture.CUSTOM_CHECKLIST_UPDATE_REQUEST;
-import static com.bang_ggood.question.CustomChecklistFixture.CUSTOM_CHECKLIST_UPDATE_REQUEST_DUPLICATED;
-import static com.bang_ggood.question.CustomChecklistFixture.CUSTOM_CHECKLIST_UPDATE_REQUEST_EMPTY;
-import static com.bang_ggood.question.CustomChecklistFixture.CUSTOM_CHECKLIST_UPDATE_REQUEST_INVALID;
-import static com.bang_ggood.user.UserFixture.USER1;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class ChecklistManageServiceTest extends IntegrationTestSupport {
 
     @Autowired
     private ChecklistManageService checklistManageService;
-    @Autowired
-    private CustomChecklistQuestionRepository customChecklistQuestionRepository;
 
     @Autowired
     private UserRepository userRepository;
@@ -59,75 +36,5 @@ class ChecklistManageServiceTest extends IntegrationTestSupport {
 
         //then
         assertThat(checklistId).isGreaterThan(0);
-    }
-
-    @DisplayName("커스텀 체크리스트 질문 조회 성공")
-    @Test
-    void readChecklistQuestions() {
-        // given
-        customChecklistQuestionRepository.saveAll(CustomChecklistFixture.CUSTOM_CHECKLIST_QUESTION_DEFAULT);
-
-        // when
-        CustomChecklistQuestionsResponse customChecklistQuestionsResponse = checklistManageService.readCustomChecklistQuestions(USER1);
-
-        // then
-        List<Integer> defaultQuestionsIds = CUSTOM_CHECKLIST_QUESTION_DEFAULT.stream()
-                .map(CustomChecklistQuestion::getQuestion)
-                .map(Question::getId)
-                .toList();
-        List<Integer> responseQuestionsIds = customChecklistQuestionsResponse.categories().stream()
-                .map(CategoryQuestionsResponse::questions)
-                .flatMap(Collection::stream)
-                .map(QuestionResponse::getQuestionId)
-                .toList();
-
-        assertThat(responseQuestionsIds).containsExactlyElementsOf(defaultQuestionsIds);
-    }
-
-    @DisplayName("커스텀 체크리스트 업데이트 성공")
-    @Test
-    void updateCustomChecklist() {
-        // given
-        CustomChecklistUpdateRequest request = CUSTOM_CHECKLIST_UPDATE_REQUEST;
-
-        // when & then
-        assertThatCode(() -> checklistManageService.updateCustomChecklist(USER1, request))
-                .doesNotThrowAnyException();
-    }
-
-    @DisplayName("커스텀 체크리스트 업데이트 실패 : 선택한 질문 개수가 0개일 때")
-    @Test
-    void updateCustomChecklist_empty_exception() {
-        // given
-        CustomChecklistUpdateRequest request = CUSTOM_CHECKLIST_UPDATE_REQUEST_EMPTY;
-
-        // when & then
-        assertThatThrownBy(() -> checklistManageService.updateCustomChecklist(USER1, request))
-                .isInstanceOf(BangggoodException.class)
-                .hasMessage(ExceptionCode.CUSTOM_CHECKLIST_QUESTION_EMPTY.getMessage());
-    }
-
-    @DisplayName("커스텀 체크리스트 업데이트 실패 : 질문이 중복될 때")
-    @Test
-    void updateCustomChecklist_duplicatedQuestion_exception() {
-        // given
-        CustomChecklistUpdateRequest request = CUSTOM_CHECKLIST_UPDATE_REQUEST_DUPLICATED;
-
-        // when & then
-        assertThatThrownBy(() -> checklistManageService.updateCustomChecklist(USER1, request))
-                .isInstanceOf(BangggoodException.class)
-                .hasMessage(ExceptionCode.QUESTION_DUPLICATED.getMessage());
-    }
-
-    @DisplayName("커스텀 체크리스트 업데이트 실패 : 질문 id가 유효하지 않을 때")
-    @Test
-    void updateCustomChecklist_invalidQuestionId_exception() {
-        // given
-        CustomChecklistUpdateRequest request = CUSTOM_CHECKLIST_UPDATE_REQUEST_INVALID;
-
-        // when & then
-        assertThatThrownBy(() -> checklistManageService.updateCustomChecklist(USER1, request))
-                .isInstanceOf(BangggoodException.class)
-                .hasMessage(ExceptionCode.QUESTION_INVALID.getMessage());
     }
 }
